// src/games/geki/types/userplaylog.types.ts
import {Types} from "mongoose";

export type GekiUserPlaylogType = {
    _id: Types.ObjectId;

    userId: string;

    sortNumber: bigint;
    placeId: number;
    placeName: string;
    playDate: Date;
    userPlayDate: Date;
    musicId: number;
    level: number;
    playKind: number;
    eventId: bigint;
    eventName: string;
    eventPoint: number;
    playedUserId1: bigint;
    playedUserId2: bigint;
    playedUserId3: bigint;
    playedUserName1: string;
    playedUserName2: string;
    playedUserName3: string;
    playedMusicLevel1: number;
    playedMusicLevel2: number;
    playedMusicLevel3: number;
    cardId1: number;
    cardId2: number;
    cardId3: number;
    cardLevel1: number;
    cardLevel2: number;
    cardLevel3: number;
    cardAttack1: number;
    cardAttack2: number;
    cardAttack3: number;
    bossCharaId: number;
    bossLevel: number;
    bossAttribute: number;
    clearStatus: number;
    techScore: number;
    techScoreRank: number;
    battleScore: number;
    battleScoreRank: number;
    platinumScore: number;
    platinumScoreStar: number;
    maxCombo: number;
    judgeMiss: number;
    judgeHit: number;
    judgeBreak: number;
    judgeCriticalBreak: number;
    rateTap: number;
    rateHold: number;
    rateFlick: number;
    rateSideTap: number;
    rateSideHold: number;
    bellCount: number;
    totalBellCount: number;
    damageCount: number;
    overDamage: number;
    isTechNewRecord: boolean;
    isBattleNewRecord: boolean;
    isOverDamageNewRecord: boolean;
    isFullCombo: boolean;
    isFullBell: boolean;
    isAllBreak: boolean;
    playerRating: number;
    battlePoint: number;
}